backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==1.29.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jiter==0.9.0
lazy-model==0.2.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
motor==3.7.0
openai==1.72.0
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pymongo==4.12.0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
